+++ scout/server/blueprints/cases/views.py
@@ -91,7 +91,16 @@ def case(institute_id, case_name):
 def sma(institute_id, case_name):
     """Visualize case SMA data - SMN CN calls"""
     institute_obj, case_obj = institute_and_case(store, institute_id, case_name)
-    data = controllers.case(store, institute_obj, case_obj)
+    data = controllers.sma_case(store, institute_obj, case_obj)
+    return dict(format="html", **data)
+
+
+@cases_bp.route("/<institute_id>/<case_name>/bionano", methods=["GET"])
+@templated("cases/case_bionano.html")
+def bionano(institute_id, case_name):
+    """Visualize case BioNano data - FSHD calls"""
+    institute_obj, case_obj = institute_and_case(store, institute_id, case_name)
+    data = controllers.bionano_case(store, institute_obj, case_obj)
     return dict(format="html", **data)